feat: Publish arm64 images as well
This adjusts the publish-dockers workflow to also publish arm64
images. It attempts to build all templates in both amd64 and arm64, but
allows the arm64 builds to fail since some of our templates
(specifically meta-reference-gpu) fail to build on arm64. After building
and pushing the architecture-specific images, it pushes a manifest list
for all the architectures we built to the default image location.

Before, we'd end up with images like `distribution-remote-vllm` that
contained only the amd64 image. Now, `distribution-remote-vllm` contains
a manifest list that contains both amd64 and arm64 images. We still
publish single manifest, architecture-specific images at
`distribution-remote-vllm-amd64` and `distribution-remote-vllm-arm64`
for really old clients, someone that specifically needs an arch-specific
image, and because it's the simplest way to run these GitHub workflows
as individual steps where we separate arch-specific image building from
manifest list publishing in GitHub Actions.

Signed-off-by: Ben Browning <bbrownin@redhat.com>